requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
psycopg2==2.9.10
djangorestframework-simplejwt==5.5.0
PyJWT==2.8.0
requests==2.32.3
